Remove duplicate helper function definitions
